Refactor FrontIntegration, create ContextSyncWrapper
    Change-type: patch
